+++ CHANGELOG.md
@@ -1,4 +1,7 @@
-# Latest changelog (v2.2.0)
+# Latest changelog (v3.0.0)
 
-- Fixed fuzzy search config property not working
+- Refactored theming system (now uses custom QML file(s) instead of JSON file(s))
+- Fixed fuzzy search config property doing nothing
+- New config properties
+- Lot of optimization
 - Removed icon and desktop shortcut